OAK-2504: oak-run debug should list a breakdown of space usage per record type - Use long to record memory usages - More space efficient way of remembering seen record ids
git-svn-id: https://svn.apache.org/repos/asf/jackrabbit/oak/trunk@1659259 13f79535-47bb-0310-9956-ffa450edef68

+++ oak-core/src/test/java/org/apache/jackrabbit/oak/plugins/segment/RecordUsageAnalyserTest.java
@@ -262,7 +262,7 @@ public class RecordUsageAnalyserTest {
     }
 
     private static void assertSizes(RecordUsageAnalyser analyser,
-            int maps, int lists, int values, int templates, int nodes) {
+            long maps, long lists, long values, long templates, long nodes) {
         assertEquals("maps sizes mismatch", maps, analyser.getMapSize());
         assertEquals("lists sizes mismatch", lists, analyser.getListSize());
         assertEquals("value sizes mismatch", values, analyser.getValueSize());
